pyproject.toml
[tool.poetry]
name = "thermostatter-api"
version = "0.1.3"
description = "thermostatter HTTP JSON API"
authors = ["WolfEYc <wolfeycode@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi-slim = "^0.111.0"
uvicorn = "^0.30.1"


[tool.poetry.scripts]
thermostatter-api = "thermostatter_api:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
